New util's function utils.bits.get_bits()
Return a chunk of the number, regarding on how many bits we need.
Also we can pass offset. It supports left and right side offset.
Right offset is recommended because we can skip some inner calculation.
But because of logic, left offset is default.

Also, updated code which use this function.


git-svn-id: http://svn.umitproject.org/svnroot/umit/branch/UMPA@3343 1105ee14-b0fa-0310-85a5-ff3eed429ff7